Handle repeated chunks efficiently. [finishes #10639513 #10639523]

+++ src/maidsafe-encrypt/self_encryption.cc
@@ -16,6 +16,7 @@
 
 #include "maidsafe-encrypt/self_encryption.h"
 
+#include <map>
 #include <set>
 #include <sstream>
 #include <vector>
@@ -118,6 +119,7 @@ int SelfEncrypt(std::istream *input_stream,
   }
 
   size_t chunk_count(chunk_sizes.size());
+  std::map<std::string, size_t> processed_chunks;
   std::string chunk_content[3], chunk_hash[3];  // sliding window of size 3
 
   // read the first 2 chunks and calculate their hashes
@@ -147,40 +149,55 @@ int SelfEncrypt(std::istream *input_stream,
       chunk_hash[idx] = data_map->chunks[(i + 2) % chunk_count].pre_hash;
     }
 
-    if (compress) {
-      chunk_content[i % 3] = crypto::Compress(chunk_content[i % 3],
-                                              kCompressionLevel);
-      if (chunk_content[i % 3].empty()) {
-        DLOG(ERROR) << "EncryptContent: Failed to compress chunk content."
-                    << std::endl;
-        return kCompressionError;
-      }
-    }
-
     ChunkDetails chunk;
     chunk.pre_hash = chunk_hash[i % 3];
     chunk.pre_size = chunk_sizes[i];
 
-    if (chunk_sizes[i] <= self_encryption_params.max_includable_chunk_size) {
-      chunk.content = chunk_content[i % 3];
-      chunk.size = chunk_content[i % 3].size();
-      // sic: chunk.hash left empty
+    if (processed_chunks.count(chunk.pre_hash) > 0) {
+      // we already processed an identical chunk before
+      ChunkDetails &prev_chunk =
+          data_map->chunks[processed_chunks[chunk.pre_hash]];
+      chunk.content = prev_chunk.content;
+      chunk.hash = prev_chunk.hash;
+      chunk.size = prev_chunk.size;
+      // DLOG(INFO) << "SelfEncrypt: chunk cache hit" << std::endl;
     } else {
-      chunk_content[i % 3] = utils::SelfEncryptChunk(chunk_content[i % 3],
-                                                     chunk_hash[(i + 1) % 3],
-                                                     chunk_hash[(i + 2) % 3]);
-
-      chunk.hash = crypto::Hash<crypto::SHA512>(chunk_content[i % 3]);
-      chunk.size = chunk_content[i % 3].size();  // encryption might add padding
-      // sic: chunk.content left empty
-
-      // write chunk file
-      fs::path chunk_path = output_dir / EncodeToHex(chunk.hash);
-      if (!utils::WriteFile(chunk_path, chunk_content[i % 3])) {
-        DLOG(ERROR) << "EncryptContent: Can't write chunk data to "
-                    << chunk_path.c_str() << std::endl;
-        return kIoError;
+      if (compress) {
+        chunk_content[i % 3] = crypto::Compress(chunk_content[i % 3],
+                                                kCompressionLevel);
+        if (chunk_content[i % 3].empty()) {
+          DLOG(ERROR) << "EncryptContent: Failed to compress chunk content."
+                      << std::endl;
+          return kCompressionError;
+        }
       }
+
+      if (chunk_sizes[i] <= self_encryption_params.max_includable_chunk_size) {
+        chunk.content = chunk_content[i % 3];
+        chunk.size = chunk_content[i % 3].size();
+        // sic: chunk.hash left empty
+      } else {
+        chunk_content[i % 3] = utils::SelfEncryptChunk(chunk_content[i % 3],
+                                                      chunk_hash[(i + 1) % 3],
+                                                      chunk_hash[(i + 2) % 3]);
+
+        chunk.hash = crypto::Hash<crypto::SHA512>(chunk_content[i % 3]);
+        chunk.size = chunk_content[i % 3].size();  // encrypt. might add padding
+        // sic: chunk.content left empty
+
+        // write chunk file
+        fs::path chunk_path = output_dir / EncodeToHex(chunk.hash);
+        if (!fs::exists(chunk_path)) {
+          // only try writing chunk if it hasn't been created before
+          if (!utils::WriteFile(chunk_path, chunk_content[i % 3])) {
+            DLOG(ERROR) << "EncryptContent: Can't write chunk data to "
+                        << chunk_path.c_str() << std::endl;
+            return kIoError;
+          }
+        }
+      }
+
+      processed_chunks[chunk.pre_hash] = data_map->chunks.size();
     }
 
     data_map->chunks.push_back(chunk);
